The HSL now works perfectly.

+++ src/Color.h
@@ -27,9 +27,15 @@ public:
   bool substract(Value* v);
   bool multiply(Value* v);
   bool divide(Value* v);
-  
-  void setHSL(int hue, double saturation,
-              double lightness);
+
+  /**
+   * The HSL to RGB conversion on
+   * http://en.wikipedia.org/wiki/HSL_and_HSV did not work at all so
+   * the calculations are from
+   * http://130.113.54.154/~monger/hsl-rgb.html, which does not list a
+   * source. 
+   */
+  void setHSL(double hue, double saturation, double lightness);
   void lighten(double percent);
   void darken(double percent);
   void saturate(double percent);
@@ -38,13 +44,49 @@ public:
   void fadeout(double percent);
   void spin(double degrees);
 
+  /**
+   * Returns the the amount of red in the color.
+   *
+   * @return an int value between 0-255
+   */
   unsigned int getRed();
+  /**
+   * Returns the the amount of green in the color.
+   *
+   * @return an int value between 0-255
+   */
   unsigned int getGreen();
+  /**
+   * Returns the the amount of blue in the color.
+   *
+   * @return an int value between 0-255
+   */
   unsigned int getBlue();
+  
+  /**
+   * Converts the internal RGB value to HSL. The source of the
+   * calculations is http://en.wikipedia.org/wiki/HSL_and_HSV except
+   * for the saturation value, which did not work.
+   */
   double* getHSL();
-  double getHue();
-  double getSaturation();
-  double getLightness();
+
+  /**
+   * Find the hue value for the color.
+   * @see getHSL();
+   */
+  int getHue();
+
+  /**
+   * Find the saturation value for the color.
+   * @see getHSL();
+   */
+  int getSaturation();
+
+  /**
+   * Find the lightness value for the color.
+   * @see getHSL();
+   */
+  int getLightness();
 };
 
 #endif
